Move Door schematic visualization to a separate class
